Get tests working with logger
