+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/api/HeterogenousFieldSupport.scala
@@ -11,7 +11,7 @@ package com.tersesystems.echopraxia.plusscala.api
  * logger.info("foo" -> Seq[Field](foo, bar)) // "foo": [foo, bar]
  * }}}
  */
-trait IterableToArrayValueImplicit { self: ValueTypeClasses =>
+trait HeterogenousFieldSupport { self: ValueTypeClasses =>
   // Render iterables as arrays (user may want to render as object, so this is broken out)
   implicit def iterableToArrayValue[V: ToValue]: ToArrayValue[Iterable[V]] = ToArrayValue.iterableToArrayValue[V]
 }
