+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/api/LoggingBase.scala
@@ -37,4 +37,3 @@ trait LoggingBase
     with FutureValueTypes
     with ToLogTypes
     with ThrowableToName
-    with IterableToArrayValueImplicit
