+++ api/src/main/scala/com/tersesystems/echopraxia/plusscala/spi/LoggerSupport.scala
@@ -1,6 +1,7 @@
 package com.tersesystems.echopraxia.plusscala.spi
 
 import com.tersesystems.echopraxia.api.FieldBuilderResult
+import com.tersesystems.echopraxia.api.Field
 import com.tersesystems.echopraxia.plusscala.api.Condition
 
 trait LoggerSupport[FB, LoggerType[_ <: FB]] { self =>
@@ -9,6 +10,8 @@ trait LoggerSupport[FB, LoggerType[_ <: FB]] { self =>
 
   def withFields(f: FB => FieldBuilderResult): LoggerType[FB]
 
+  //def withFields(fields: => Seq[Field]): LoggerType[FB]
+  
   def withThreadContext: LoggerType[FB]
 
   def withFieldBuilder[T <: FB](newBuilder: T): LoggerType[T]
