+++ logger/src/test/scala/com/tersesystems/echopraxia/plusscala/Main.scala
@@ -1,145 +1,100 @@
 package com.tersesystems.echopraxia.plusscala
 
-import com.tersesystems.echopraxia.api.{Field, Level}
-import com.tersesystems.echopraxia.spi.{CoreLoggerFactory, DefaultField, PresentationHintAttributes}
-import com.tersesystems.echopraxia.plusscala.api._
-import com.tersesystems.echopraxia.plusscala.spi.Utils
-import enumeratum.EnumEntry._
-import enumeratum._
+import com.tersesystems.echopraxia.api.Field
 
-import java.time.Instant
+import com.tersesystems.echopraxia.plusscala.api.HeterogenousFieldSupport
+
+import java.util.Currency
+import java.util.UUID
+import scala.concurrent.Future
 
 object Main {
-
-  sealed trait Names extends EnumEntry with Snakecase
-
-  object Names extends Enum[Names] {
-
-    /*
-     `findValues` is a protected method that invokes a macro to find all `Greeting` object declarations inside an `Enum`
-
-     You use it to implement the `val values` member
-     */
-    val values = findValues
-
-    case object ExpirationDate extends Names
-
-    case object Number     extends Names
-    case object CreditCard extends Names
-
-    case object Book extends Names
-
-    case object Instant  extends Names
-    case object Title    extends Names
-    case object Author   extends Names
-    case object Category extends Names
-    case object Price    extends Names
-
-    case object Store extends Names
-  }
-
-  trait CreditCardFieldBuilder extends ArgsFieldBuilder with HasName {
-    override type FieldType = DefaultField
-    protected val fieldClass: Class[DefaultField] = classOf[DefaultField]
-
-    import Names._
-    type Name = Names
-
-    type Id[A] = A
-    implicit def narrow[T <: Singleton](t: T): Id[T] = t
-
-    override def keyValue[V: ToValue: ToValueAttributes](key: Name, value: V) = {
-      val ev = implicitly[ToValueAttributes[V]]
-      val attributes = ev.toAttributes(ev.toValue(value))
-      Utils.newField(key.entryName, ToValue(value), attributes, fieldClass)
-    }
-
-    // ------------------------------------------------------------------
-    // value
-
-    override def value[V: ToValue: ToValueAttributes](key: Name, value: V) = {
-      val ev = implicitly[ToValueAttributes[V]]
-      val attributes = ev.toAttributes(ev.toValue(value)).plus(PresentationHintAttributes.asValueOnly())
-      Utils.newField(key.entryName, ToValue(value), attributes, fieldClass)
-    }
-
-    implicit def creditCardToValue(implicit cap: Sensitive = Censored): ToValue[CreditCard] = cc => {
-      ToObjectValue(
-        sensitiveKeyValue(Number, cc.number),
-        keyValue(ExpirationDate, cc.expirationDate)
-      )
-    }
-
-    def sensitiveKeyValue(name: Names, value: String)(implicit cap: Sensitive = Censored): Field = {
-      cap match {
-        case Censored =>
-          keyValue(name, "[CENSORED]")
-        case Explicit =>
-          keyValue(name, value)
-      }
-    }
-  }
-
-  trait MyFieldBuilder extends CreditCardFieldBuilder {
-    override protected val fieldClass: Class[DefaultField] = classOf[DefaultField]
-
-    implicit val instantToStringValue: ToValue[Instant] = (t: Instant) => ToValue(t.toString)
-
-    implicit val bookToObjectValue: ToObjectValue[Book] = { book =>
-      ToObjectValue(
-        keyValue(Names.Category, book.category),
-        keyValue(Names.Author, book.author),
-        keyValue(Names.Title, book.title),
-        keyValue(Names.Price, book.price)
-      )
-    }
-
-    implicit def mapToObjectValue[N <: Name, V: ToValue]: ToObjectValue[Map[N, V]] = m =>
-      ToObjectValue(m.map { case (k, v) =>
-        keyValue(k, v)
-      })
-  }
-
-  object MyFieldBuilder extends MyFieldBuilder
-
-  private val logger = Logger(CoreLoggerFactory.getLogger("", getClass), MyFieldBuilder)
-
-  case class Book(category: String, author: String, title: String, price: Double)
-
-  case class CreditCard(number: String, expirationDate: String)
-
   def main(args: Array[String]): Unit = {
-
-    implicit val sensitive = Explicit
-    val creditCard         = CreditCard("4111 1111 1111 1111", "04/23")
-    logger.info("{}", _.keyValue(Names.CreditCard, creditCard))
-
-    val refBook = Book("reference", "Nigel Rees", "Sayings of the Century", 8.95)
-    logger.info(
-      "{}",
-      fb => {
-        import fb._
-        fb.obj(Names.Store, fb.array(Names.Book, Seq(refBook)))
-      }
-    )
-
-    logger.info("{}", _.keyValue(Names.Book, refBook))
-
-    logger.info("{}", _.keyValue(Names.Instant, Instant.now()))
-
-    logger.debug(infoOrHigherCondition, "INFO message")
+    val printer = new Printer()
+    printer.print()
   }
+}
 
-  val infoOrHigherCondition: Condition = Condition((level, _) => level >= Level.INFO)
 
-  val fooCondition: Condition = Condition(_.fields.exists(_.name == "foo"))
+class Printer extends Logging with HeterogenousFieldSupport {
+  private val logger = LoggerFactory.getLogger(getClass, MyFieldBuilder)
 
-  val infoAndFoo: Condition = infoOrHigherCondition and fooCondition
+  private val USD = Currency.getInstance("USD")
 
-  sealed trait Sensitive
+  def print(): Unit = {
+    val person1 = Person("Person1", 12)
+    val person2 = Person("Person2", 15)
 
-  case object Censored extends Sensitive
+    // This uses the "personToField" mapping
+    logger.info("template shows {}", person1)
+    logger.info(person1)
 
-  case object Explicit extends Sensitive
+    // Can define custom mapping with tuples
+    logger.info("person1" -> person1)
+    logger.info("person1" -> person1, "person2" -> person2)
 
+    // Options work out of the box
+    val optPerson: Option[Person] = Option(person1)
+    logger.info("optPerson" -> optPerson)
+
+    // XXX this doesn't work in 2.12 but works in 2.13 and 3
+    // logger.info("optPerson" -> None)
+
+    // As does either
+    logger.info("eitherPerson" -> Left(person1))
+
+    // And so do lists
+    logger.info("people" -> Seq(person1, person2))
+
+    // and maps (defined in Logging.scala)
+    logger.info("people" -> Map("person1" -> person1, "person2" -> person2))
+
+    // And even tuples  (defined in Logging.scala)
+    // XXX This doesn't work in 2.12, works in 2.13 and 3
+    logger.info("intToPersonMap" -> (1 -> person1))
+
+    // support for exceptions
+    logger.error(new IllegalStateException())
+
+    // Log with conditional logging
+    if (logger.isInfoEnabled) {
+      logger.info("p1" -> person1, "p2" -> person2, "p3" -> person1)
+    }
+
+    // Complex objects are no problem
+    val book1 = Book(
+      Category("reference"),
+      Author("Nigel Rees"),
+      Title("Sayings of the Century"),
+      Price(amount = 8.95, currency = Currency.getInstance("USD"))
+    )
+    logger.info(book1)
+
+    // You can also render case classes using custom presentation while using
+    // structured JSON, i.e. this will render $8.95 in pattern layout
+    // while still rendering as a JSON object
+    logger.info(Price(8.95, USD))
+
+    val creditCard = CreditCard("4111 1111 1111 1111", "04/23")
+    logger.info("{}", creditCard)
+
+    // If you want to render fields as an object, you can use ToObjectValue
+    logger.info("object" -> ToObjectValue(book1, person1)) // object={book={}, person={}}
+
+    // For heterogeneous fields you'll need to use `Seq[Field]` explicitly, or use info.v as seen below
+    logger.info("object" -> Seq[Field](book1, person1)) // object=[book={}, person={}]
+
+    // For heterogeneous values you'll want to specify Seq[Value[_]] to give implicit conversion some clues
+    logger.info("oneTrueString" -> Seq(ToValue(1), ToValue(true), ToValue("string")))
+
+    // You can also use "withFields" to render JSON on every message (this will not show in line format)
+    //logger.withFields(Seq(book1, person1)).info("testing")
+
+    // Can also log using class name
+    logger.info(UUID.randomUUID)
+
+    // Logging futures is also possible, and can include names (defined in Logging.scala)
+    logger.info(Future.successful(true))
+    logger.info(Future.successful("String"))
+  }
 }
